Another approach at Maze performance
This time, store only the north and west passages at each cell.
That way, there's only one bit to flip for each setPassage.  This
improves constructor performance further, and triples setPassage
performance.

It slows down getPassage (though not as much as before the previous
performance commit), and cell, but this should help more with
maze generation.